Move repayment API calls into dedicated module
