+++ src/components/repayments/Repayments.jsx
@@ -6,7 +6,8 @@ import { useTranslation } from "react-i18next";
 import LoanDetails from "../dashboard/LoanDetails";
 import RepaymentsList from "../dashboard/RepaymentsList";
 import { useState, useEffect } from "react";
-import { fetchRepayments, getLoanDetails } from "../../api/apiData";
+import { fetchRepayments } from "../../api/repaymentsApi";
+import { getLoanDetails } from "../../api/apiData";
 import calendarIcon from "../../assets/calendar icon.svg";
 import searchIcon from "../../assets/search icon.svg";
 import { useNavigate, Link } from "react-router-dom";
